requirements.txt
openai==0.28.1
slack_bolt==1.23.0
slack_sdk==3.35.0
faiss-cpu==1.11.0
python-dotenv==1.1.1
pdfplumber==0.10.4
tqdm==4.67.1
numpy>=1.24.0
Pillow>=9.0.0
pytesseract>=0.3.10

